+++ Cargo.toml
@@ -65,6 +65,9 @@ solders-rpc-filter = { path = "./crates/rpc-filter", version = "0.14.4" }
 solders-rpc-config = { path = "./crates/rpc-config", version = "0.14.4" }
 solders-rpc-common = { path = "./crates/rpc-common", version = "0.14.4" }
 solders-rpc-errors = { path = "./crates/rpc-errors", version = "0.14.4" }
+solders-rpc-errors-common = { path = "./crates/rpc-errors-common", version = "0.14.4" }
+solders-rpc-errors-tx-status = { path = "./crates/rpc-errors-tx-status", version = "0.14.4" }
+solders-rpc-errors-no-tx-status = { path = "./crates/rpc-errors-no-tx-status", version = "0.14.4" }
 solders-rpc-requests = { path = "./crates/rpc-requests", version = "0.14.4" }
 solders-rpc-responses = { path = "./crates/rpc-responses", version = "0.14.4" }
 solders-rpc-responses-common = { path = "./crates/rpc-responses-common", version = "0.14.4" }
