+++ crates/rpc-errors-common/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "solders-rpc-errors-common"
+version = "0.14.4"
+edition = "2021"
+include = ["/src"]
+license = "MIT"
+repository = "https://github.com/kevinheavey/solders"
