+++ crates/rpc-errors-no-tx-status/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solders-rpc-errors"
+name = "solders-rpc-errors-no-tx-status"
 version = "0.14.4"
 edition = "2021"
 include = ["/src"]
@@ -9,6 +9,8 @@ repository = "https://github.com/kevinheavey/solders"
 [dependencies]
 pyo3 = { workspace = true }
 serde = { workspace = true }
+derive_more = { workspace = true }
 solders-macros = { workspace = true }
-solders-rpc-errors-tx-status = { workspace = true }
-solders-rpc-errors-no-tx-status = { workspace = true }
+solders-traits-core = { workspace = true }
+solders-transaction-error = { workspace = true }
+solders-rpc-errors-common = { workspace = true }
