+++ crates/rpc-errors-tx-status/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "solders-rpc-errors-tx-status"
+version = "0.14.4"
+edition = "2021"
+include = ["/src"]
+license = "MIT"
+repository = "https://github.com/kevinheavey/solders"
+
+[dependencies]
+pyo3 = { workspace = true }
+serde = { workspace = true }
+derive_more = { workspace = true }
+solders-macros = { workspace = true }
+solders-traits-core = { workspace = true }
+solders-transaction-error = { workspace = true }
+solders-rpc-errors-common = { workspace = true }
+solders-rpc-common = { workspace = true }
