+++ crates/rpc-responses/Cargo.toml
@@ -25,7 +25,8 @@ solders-rpc-responses-common = { workspace = true }
 solders-rpc-responses-tx-status = { workspace = true }
 solders-rpc-response-data-boilerplate = { workspace = true }
 solders-rpc-common = { workspace = true }
-solders-rpc-errors = { workspace = true }
+solders-rpc-errors-no-tx-status = { workspace = true }
+solders-rpc-errors-tx-status = { workspace = true }
 solders-rpc-version = { workspace = true }
 solders-pubkey = { workspace = true }
 solders-signature = { workspace = true }
